requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cloudpickle==1.2.1
contextlib2==0.5.5
cryptography==2.7
cycler==0.10.0
Cython==0.29.12
cytoolz==0.10.0
dask==2.2.0
decorator==4.4.0
glog==0.3.1
graphviz==0.11.1
h5py==2.10.0
idna==2.8
imageio==2.5.0
jiwer==1.3.2
kiwisolver==1.1.0
lmdb==0.97
matplotlib==3.1.0
mkl-fft==1.0.14
mkl-random==1.0.2
mkl-service==2.0.2
munch==2.3.2
networkx==2.3
numpy==1.16.4
olefile==0.46
opencv-python==4.1.0.25
opt-einsum==2.3.2
pandas==0.25.1
Pillow==6.1.0
pretrainedmodels==0.7.4
pycparser==2.19
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyro-ppl==0.3.4
PySocks==1.7.0
python-dateutil==2.8.0
python-gflags==3.1.2
pytz==2019.2
PyWavelets==1.0.3
requests==2.22.0
scikit-image==0.15.0
scipy==1.3.0
six==1.12.0
toolz==0.10.0
torch==1.1.0
torchsummary==1.5.1
torchsummaryX==1.3.0
torchvision==0.3.0
tornado==6.0.3
tqdm==4.32.2
urllib3==1.25.3
